Temporal Fix to SIRT weigth
Problem is that the way the W is computed is wrong. It would work fine for fake data, but it wont for real data, as whatever is in the projector is not only whatever is in the image we want to recosntruct, often there are other thigns that we leave out of the image region. The way W was computed would assume that everything in D should be generated in I and that creates HUGE overstimateion og HU units on the cornres on the image. SIRT now has a temporary fix that we need to test/modify, and still needs fixing in all the rest of algorithms.

See also: #12